+++ docs/changelog.md
@@ -3,6 +3,7 @@
 ## [v0.3.3](https://github.com/0x2142/frigate-notify/releases/tag/v0.3.3) - Upcoming Release
  - Additional filtering for alerts:
      - Sublabels
+     - Minimum label score
 
 ## [v0.3.2](https://github.com/0x2142/frigate-notify/releases/tag/v0.3.2) - Jun 13 2024